ui/rsl: use common load level input
