+++ ui/rsl/src/components/measures/TripAndLoadInput.tsx
@@ -1,0 +1,47 @@
+import React, { ReactNode } from "react";
+
+import { TripServiceInfo } from "@/api/protocol/motis";
+import { LoadLevel } from "@/api/protocol/motis/paxforecast";
+
+import TripPicker from "@/components/TripPicker";
+import LoadInput, { allLoadLevels } from "@/components/measures/LoadInput";
+
+export type TripAndLoadInputProps = {
+  selectedTrip: TripServiceInfo | undefined;
+  selectedLevel: LoadLevel;
+  onTripSelected: (tsi: TripServiceInfo | undefined) => void;
+  onLevelSelected: (level: LoadLevel) => void;
+  loadLevels?: LoadLevel[];
+  children?: ReactNode;
+};
+
+function TripAndLoadInput({
+  selectedTrip,
+  selectedLevel,
+  onTripSelected,
+  onLevelSelected,
+  loadLevels = allLoadLevels,
+  children,
+}: TripAndLoadInputProps) {
+  return (
+    <div className="flex justify-between items-center gap-2">
+      <TripPicker
+        onTripPicked={onTripSelected}
+        clearOnPick={false}
+        longDistanceOnly={false}
+        initialTrip={selectedTrip}
+        key={JSON.stringify(selectedTrip)}
+        className="w-32 flex-shrink-0"
+      />
+      <LoadInput
+        loadLevels={loadLevels}
+        selectedLevel={selectedLevel}
+        onLevelSelected={onLevelSelected}
+        className="flex-grow"
+      />
+      {children}
+    </div>
+  );
+}
+
+export default TripAndLoadInput;
